Part 2 – Child classes
In fact, the work for Part 2 has already been completed in the previous upload update, and this upload is only for overall smoothness

2. Added the GeneralPractitioner subclass, inheriting all instance variables and methods of HealthProfessional
Unique instance variable: ReferralRecord, used to represent the referral records of general practitioners.
﻿
3. Added Specialist subclass, inheriting all instance variables and methods of HealthProfessional
Unique instance variables: specializeArea and patientLimit, used to represent the expert's professional field and the upper limit of patient reception.